[credentialless]: Fix WPT reliability.
It sounds like some server of the official web-test-platform runner can
drop some requests when too many are sent in parallel.

This patch ensures both send/receive function retry as long as there is
an error. We also limit the number of concurrent access to the fetch API
to one per document.

Bug: 1196073
Change-Id: I2c2ef6c2b7c6d7bbeeea2709f1c6bacbcb081419